Add intro.py file for intro function
