+++ intro.py
@@ -1,0 +1,9 @@
+def intro():  # Improves Interface
+    print(""" 
+░█▀▄░█▀█░█▀█░█░█        
+░█▀▄░█░█░█░█░█▀▄        
+░▀▀░░▀▀▀░▀▀▀░▀░▀        
+░█▀▀░█▀▀░█▀█░█▀▄░█▀▀░█░█
+░▀▀█░█▀▀░█▀█░█▀▄░█░░░█▀█
+░▀▀▀░▀▀▀░▀░▀░▀░▀░▀▀▀░▀░▀                                                                        
+    """)
